Refactored scss
