+++ src/_scss/pages/awardV2/awardFinancialAssistance.scss
@@ -1,6 +1,4 @@
 /* Styles specific to Financial Assistance */
 
 .award-financial-assistance {
-    @import "./overview/_amountDates";
-    @import "./overview/_awardRecipient";
 }
