+++ src/_scss/pages/awardV2/awardPage.scss
@@ -20,35 +20,15 @@
                     @include display(flex);
                     @include flex(0 0 auto);
                     font-weight: $font-semibold;
+                    padding-right: rem(20);
                 }
-                .award__heading-glossary {
-                    @include display(flex);
-                    @include flex(0 0 auto);
-                    padding: 0 rem(20) 0 rem(9);
-                    a {
-                        width: rem(19);
-                        height: rem(25);
-                        svg {
-                            width: rem(19);
-                            height: rem(25);
-                            fill: $color-gray-light;
-                        }
-                        &:hover {
-                            cursor: pointer;
-                            svg {
-                                fill: $color-link;
-                            }
-                        }
-                    }
-                }
-                .award__heading-id {
-                    font-weight: $font-light;
-                }
-
                 .award__heading-lable {
                     font-weight: $font-normal;
                     display: inline-block;
-                    margin-right: rem(5);
+                    padding-right: rem(10);
+                }
+                .award__heading-id {
+                    font-weight: $font-light;
                 }
             }
 
@@ -73,6 +53,7 @@
                         padding-left: rem(15);
                     }
                 }
+
                 @import "./visualizations/awardViz";
 
                 .text-button-container__button {
@@ -90,24 +71,10 @@
                    font-weight: $font-light;
                 }
             }
-
-            .text-button-container__button {
-                background-color: transparent;
-                color: $color-cool-blue;
-                padding: 0;
-                margin: 0;
-            }
-
-            hr {
-                opacity: 0.5;
-            }
-
-            .data-values {
-               font-weight: $font-light;
-            }
-
         }
 
+        @import "./visualizations/overview/dates";
+        @import "./visualizations/overview/agencyRecipient";
         @import "./awardContract";
         @import "./awardFinancialAssistance";
     }
