+++ src/js/components/awardv2/visualizations/overview/AgencyRecipient.jsx
@@ -1,5 +1,5 @@
 /**
- * AwardRecipient.jsx
+ * AgencyRecipient.jsx
  * Created by David Trinh 10/11/2018
  **/
 
@@ -14,13 +14,13 @@ const propTypes = {
     jumpToSection: PropTypes.func
 };
 
-export default class AwardRecipient extends React.Component {
+export default class AgencyRecipient extends React.Component {
     render() {
         return (
-            <div className="award__col award-recipient">
-                <div className="award-recipient__awarding">
-                    <span className="award-recipient__awarding-title">Awarding Agency</span>
-                    <div className="award-recipient__awarding-detail">
+            <div className="award__col agency-recipient">
+                <div className="agency-recipient__awarding">
+                    <span className="agency-recipient__awarding-title">Awarding Agency</span>
+                    <div className="agency-recipient__awarding-detail">
                         <a href={`/#/agency/${this.props.awardingAgency.id}`}>
                             {`${this.props.awardingAgency.toptierName} ${this.props.awardingAgency.toptierAbbr}`}
                         </a>
@@ -28,19 +28,19 @@ export default class AwardRecipient extends React.Component {
                     <br />
                     <div>
                         <span
-                            className={`award-recipient__awarding-icon ${this.props.category !== 'contract' ? 'award-recipient__awarding-icon_hide' : ''}`}>
+                            className={`agency-recipient__awarding-icon ${this.props.category !== 'contract' ? 'agency-recipient__awarding-icon_hide' : ''}`}>
                             <Icons.ArrowDown />
                         </span>
                         <button
-                            className={`award-recipient__awarding-info ${this.props.category !== 'contract' ? 'award-recipinet__awarding-info_hide' : ''}`}
+                            className={`agency-recipient__awarding-info ${this.props.category !== 'contract' ? 'award-recipinet__awarding-info_hide' : ''}`}
                             onClick={() => this.props.jumpToSection('additional-information')}>
                             View additional information
                         </button>
                     </div>
                 </div>
-                <div className="award-recipient__recipient">
-                    <span className="award-recipient__recipient-title">Recipient</span> <br />
-                    <span className="award-recipient__recipient-detail">
+                <div className="agency-recipient__recipient">
+                    <span className="agency-recipient__recipient-title">Recipient</span> <br />
+                    <span className="agency-recipient__recipient-detail">
                         <a href={`/#/recipient/${this.props.recipient.internalId}`}>
                             {this.props.recipient.name}
                         </a>
@@ -50,4 +50,4 @@ export default class AwardRecipient extends React.Component {
         );
     }
 }
-AwardRecipient.propTypes = propTypes;
+AgencyRecipient.propTypes = propTypes;
